Migrate LSP config from deprecated lspconfig API to vim.lsp.config
The require('lspconfig') framework is deprecated in nvim-lspconfig and
will be removed in v3.0.0. This migrates to the new vim.lsp.config API
introduced in Neovim 0.11, which provides the same functionality through
the native Neovim LSP interface.

Changes:
- Replace require('lspconfig')[name].setup() with vim.lsp.config[name]
- Add explicit vim.lsp.enable(name) call to enable configured servers
- Remove the lspconfig require statement

This eliminates the deprecation warning while maintaining identical
functionality for all LSP servers.