Fix worktree script to use correct project name when called from within a worktree
When creating or removing worktrees from within an existing worktree, the script
incorrectly used the current worktree directory name as the project name instead
of finding the main repository location.

For example, running 'worktree create new-feature' from within
~/worktrees/myproject/feature-branch would incorrectly try to create
~/worktrees/feature-branch/new-feature instead of ~/worktrees/myproject/new-feature.

The fix adds get_main_worktree() which uses 'git worktree list --porcelain' to find
the main repository (always the first entry in the list), then derives the project
name from that instead of the current working directory.

This fix is applied to:
- create command
- remove command
- list command
- complete_worktree_names function (shell completion)